+++ include/ResourceManager.h
@@ -53,39 +53,39 @@ class ResourceManager
 		ResourceManager();
 		~ResourceManager();
 
-		kShortestPathReturn* calculate_SP_path(unsigned short int src, unsigned short int dest, unsigned short int k, unsigned short int ci);
-		kShortestPathReturn* calculate_LORA_path(unsigned short int src, unsigned short int dest, unsigned short int k, unsigned short int ci);
-		kShortestPathReturn* calculate_PAR_path(unsigned short int src_index, unsigned short int dest_index, unsigned short int k, unsigned short int ci);
-		kShortestPathReturn* calculate_IA_path(unsigned short int src_index, unsigned short int dest_index, unsigned short int ci);
-		kShortestPathReturn* calculate_QM_path(unsigned short int src_index, unsigned short int dest_index, unsigned short int k, unsigned short int ci);
-		kShortestPathReturn* calculate_AQoS_path(unsigned short int src_index, unsigned short int dest_index, unsigned short int k, unsigned short int ci);
-		kShortestPathReturn* calculate_DP_path(unsigned short int src_index, unsigned short int dest_index, unsigned short int k, unsigned short int ci);
-		kShortestPathReturn* calculate_ACO_path(unsigned short int src_index, unsigned short int dest_index, unsigned short int k, unsigned short int ci);
-		kShortestPathReturn* calculate_MM_ACO_path(unsigned short int src_index, unsigned short int dest_index, unsigned short int k, unsigned short int ci);
+		kShortestPathReturn* calculate_SP_path(unsigned int src, unsigned int dest, unsigned int k, unsigned int ci);
+		kShortestPathReturn* calculate_LORA_path(unsigned int src, unsigned int dest, unsigned int k, unsigned int ci);
+		kShortestPathReturn* calculate_PAR_path(unsigned int src_index, unsigned int dest_index, unsigned int k, unsigned int ci);
+		kShortestPathReturn* calculate_IA_path(unsigned int src_index, unsigned int dest_index, unsigned int ci);
+		kShortestPathReturn* calculate_QM_path(unsigned int src_index, unsigned int dest_index, unsigned int k, unsigned int ci);
+		kShortestPathReturn* calculate_AQoS_path(unsigned int src_index, unsigned int dest_index, unsigned int k, unsigned int ci);
+		kShortestPathReturn* calculate_DP_path(unsigned int src_index, unsigned int dest_index, unsigned int k, unsigned int ci);
+		kShortestPathReturn* calculate_ACO_path(unsigned int src_index, unsigned int dest_index, unsigned int k, unsigned int ci);
+		kShortestPathReturn* calculate_MM_ACO_path(unsigned int src_index, unsigned int dest_index, unsigned int k, unsigned int ci);
 
-		int choose_wavelength(CreateConnectionProbeEvent* ccpe, unsigned short int ci);
+		int choose_wavelength(CreateConnectionProbeEvent* ccpe, unsigned int ci);
 
-		double estimate_Q(short int lambda, Edge **Path, unsigned short int pathLen, double *xpm, double *fwm, double *ase, unsigned short int ci);
+		double estimate_Q(int lambda, Edge **Path, unsigned int pathLen, double *xpm, double *fwm, double *ase, unsigned int ci);
 
 		void initSPMatrix();
 		void freeSPMatrix();
 
 		double path_fwm_term(int spans,double fi,double fj, double fk,double fc,int dgen);
-		double path_xpm_term(short int spans, short int lambda, short int wave);
+		double path_xpm_term(int spans, int lambda, int wave);
 
-		void print_connection_info(CreateConnectionProbeEvent* ccpe, double Q_factor, double ase, double fwm, double xpm, unsigned short int ci);
+		void print_connection_info(CreateConnectionProbeEvent* ccpe, double Q_factor, double ase, double fwm, double xpm, unsigned int ci);
 
 		double* sys_fs;
 		vector<int>* fwm_combinations;
 
-		unsigned short int* span_distance;
+		unsigned int* span_distance;
 
 	private:
-		double path_ase_noise(short int lambda, Edge **Path, unsigned short int pathLen, unsigned short int ci);
+		double path_ase_noise(int lambda, Edge **Path, unsigned int pathLen, unsigned int ci);
 
-		double path_fwm_noise(short int lambda, Edge **Path, unsigned short int pathLen, unsigned short int ci);
+		double path_fwm_noise(int lambda, Edge **Path, unsigned int pathLen, unsigned int ci);
 
-		double path_xpm_noise(short int lambda, Edge **Path, unsigned short int pathLen, unsigned short int ci);
+		double path_xpm_noise(int lambda, Edge **Path, unsigned int pathLen, unsigned int ci);
 		
 		int build_FWM_fs(double *inter_fs,int *inter_indecies, int lambda);
 		int wave_combines(double fc, double *fs,int fs_num, vector<int> &fs_coms);
@@ -95,22 +95,22 @@ class ResourceManager
 		double* sys_link_xpm_database;
 		int sys_fs_num;
 
-		int first_fit(CreateConnectionProbeEvent* ccpe, unsigned short int ci, bool* wave_available);
-		int first_fit_with_ordering(CreateConnectionProbeEvent* ccpe, unsigned short int ci, bool* wave_available);
+		int first_fit(CreateConnectionProbeEvent* ccpe, unsigned int ci, bool* wave_available);
+		int first_fit_with_ordering(CreateConnectionProbeEvent* ccpe, unsigned int ci, bool* wave_available);
 
-		int random_fit(CreateConnectionProbeEvent* ccpe, unsigned short int ci, bool* wave_available,unsigned short int numberAvailableWaves);
+		int random_fit(CreateConnectionProbeEvent* ccpe, unsigned int ci, bool* wave_available,unsigned int numberAvailableWaves);
 
-		int most_used(CreateConnectionProbeEvent* ccpe, unsigned short int ci, bool* wave_available);
+		int most_used(CreateConnectionProbeEvent* ccpe, unsigned int ci, bool* wave_available);
 
-		int quality_first_fit(CreateConnectionProbeEvent* ccpe, unsigned short int ci, bool* wave_available,unsigned short int numberAvailableWaves);
-		int quality_first_fit_with_ordering(CreateConnectionProbeEvent* ccpe, unsigned short int ci, bool* wave_available,unsigned short int numberAvailableWaves);
+		int quality_first_fit(CreateConnectionProbeEvent* ccpe, unsigned int ci, bool* wave_available,unsigned int numberAvailableWaves);
+		int quality_first_fit_with_ordering(CreateConnectionProbeEvent* ccpe, unsigned int ci, bool* wave_available,unsigned int numberAvailableWaves);
 
-		int quality_random_fit(CreateConnectionProbeEvent* ccpe, unsigned short int ci, bool* wave_available,unsigned short int numberAvailableWaves);
+		int quality_random_fit(CreateConnectionProbeEvent* ccpe, unsigned int ci, bool* wave_available,unsigned int numberAvailableWaves);
 
-		int quality_most_used(CreateConnectionProbeEvent* ccpe, unsigned short int ci, bool* wave_available,unsigned short int numberAvailableWaves);
+		int quality_most_used(CreateConnectionProbeEvent* ccpe, unsigned int ci, bool* wave_available,unsigned int numberAvailableWaves);
 
-		int least_quality_fit(CreateConnectionProbeEvent* ccpe, unsigned short int ci, bool* wave_available);
-		int most_quality_fit(CreateConnectionProbeEvent* ccpe, unsigned short int ci, bool* wave_available);
+		int least_quality_fit(CreateConnectionProbeEvent* ccpe, unsigned int ci, bool* wave_available);
+		int most_quality_fit(CreateConnectionProbeEvent* ccpe, unsigned int ci, bool* wave_available);
 
 		void precompute_fwm_fs(vector<int> &fwm_nums);
 		void precompute_fwm_combinations();
@@ -125,14 +125,14 @@ class ResourceManager
 		kShortestPathEdges* kSP_edgeList;
 
 		void calc_min_spans();
-		unsigned short int calculate_span_distance(unsigned short int src, unsigned short int dest);
+		unsigned int calculate_span_distance(unsigned int src, unsigned int dest);
 
-		short int* wave_ordering;
+		int* wave_ordering;
 
 		void generateWaveOrdering();
 
-		short int getLowerBound(short int w, short int n);
-		short int getUpperBound(short int w, short int n);
+		int getLowerBound(int w, int n);
+		int getUpperBound(int w, int n);
 };
 
 struct Ant
