+++ kshortestpath/src/MainP.cpp
@@ -85,12 +85,12 @@ void copyResults(std::vector<CQYDirectedPath*> &topK_shortest_paths, const kShor
 		}
 	}
 
-	for(unsigned short int a = 0; a < topK_shortest_paths.size(); ++a)
+	for(unsigned int a = 0; a < topK_shortest_paths.size(); ++a)
 	{
 		retVal->pathcost[a] = float(topK_shortest_paths[a]->GetCost());
 		retVal->pathlen[a] = topK_shortest_paths[a]->GetLength();
 
-		for(unsigned short int b = 0; b < retVal->pathlen[a]; ++b)
+		for(unsigned int b = 0; b < retVal->pathlen[a]; ++b)
 		{
 			retVal->pathinfo[a * (params.total_nodes - 1) + b] = topK_shortest_paths[a]->GetVertexList()[b];
 		}
